Fix NaN loss caused by float16 gradient overflow during backbone training
ModelRegistry: load backbone in float32 when trainable=True. float16 has
a max representable value of ~65504; gradients through transformer layers
overflow to inf/NaN during backward without gradient scaling. Frozen models
keep float16 (inference-only, no backward pass).

RelationEncoder: cast last_hidden_state to float32 immediately after the
backbone forward. The hidden states were float16 while the one-hot masks
were float32; PyTorch kept the product in float16, creating a second
overflow path independent of whether the backbone itself is trainable.

+++ deepref/encoder/relation_encoder.py
@@ -58,7 +58,7 @@ class RelationEncoder(LLMEncoder):
             (B, 3H) concatenated e1 / e2 / mask hidden states
         """
         outputs = self.registry.run_from_input_ids(self.model_name, token, attention_mask=att_mask)
-        hidden = outputs.last_hidden_state  # (B, L, H)
+        hidden = outputs.last_hidden_state.float()  # (B, L, H) — cast to float32
 
         onehot_e1   = torch.zeros(hidden.size()[:2]).float().to(hidden.device)  # (B, L)
         onehot_e2   = torch.zeros(hidden.size()[:2]).float().to(hidden.device)  # (B, L)
